requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# CLIP specific
openai-clip>=1.0.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
click>=8.1.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Evaluation & visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0